pyproject.toml
[tool.poetry]
name = "notebook-email-verification"
version = "0.1.0"
description = ""
authors = ["Volodymyr Kurov <vokur13@gmail.com>"]
readme = "README.md"
packages = [{include = "notebook_email_verification"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.21"
psycopg2-binary = "^2.9.7"
libgravatar = "^1.0.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
alembic = "^1.12.0"
email-validator = "^2.0.0.post2"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
redis = "^5.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
